Add optional settings for WithMappingFromOpenApi
