+++ src/WireMock.Net.OpenApiParser/Extensions/WireMockServerExtensions.cs
@@ -1,12 +1,16 @@
-﻿using System;
-using System.IO;
+﻿using System.IO;
 using System.Linq;
 using Microsoft.OpenApi.Models;
 using Microsoft.OpenApi.Readers;
+using Stef.Validation;
+using WireMock.Net.OpenApiParser.Settings;
 using WireMock.Server;
 
 namespace WireMock.Net.OpenApiParser.Extensions
 {
+    /// <summary>
+    /// Some extension methods for <see cref="IWireMockServer"/>.
+    /// </summary>
     public static class WireMockServerExtensions
     {
         /// <summary>
@@ -17,16 +21,22 @@ namespace WireMock.Net.OpenApiParser.Extensions
         /// <param name="diagnostic">Returns diagnostic object containing errors detected during parsing</param>
         public static IWireMockServer WithMappingFromOpenApiFile(this IWireMockServer server, string path, out OpenApiDiagnostic diagnostic)
         {
-            if (server == null)
-            {
-                throw new ArgumentNullException(nameof(server));
-            }
-            if (string.IsNullOrEmpty(path))
-            {
-                throw new ArgumentNullException(nameof(path));
-            }
+            return WithMappingFromOpenApiFile(server, path, null, out diagnostic);
+        }
 
-            var mappings = new WireMockOpenApiParser().FromFile(path, out diagnostic);
+        /// <summary>
+        /// Register the mappings via an OpenAPI (swagger) V2 or V3 file.
+        /// </summary>
+        /// <param name="server">The WireMockServer instance</param>
+        /// <param name="path">Path containing OpenAPI file to parse and use the mappings.</param>
+        /// <param name="diagnostic">Returns diagnostic object containing errors detected during parsing</param>
+        /// <param name="settings">Additional settings</param>
+        public static IWireMockServer WithMappingFromOpenApiFile(this IWireMockServer server, string path, WireMockOpenApiParserSettings settings, out OpenApiDiagnostic diagnostic)
+        {
+            Guard.NotNull(server, nameof(server));
+            Guard.NotNullOrEmpty(path, nameof(path));
+
+            var mappings = new WireMockOpenApiParser().FromFile(path, settings, out diagnostic);
 
             return server.WithMapping(mappings.ToArray());
         }
@@ -41,6 +51,20 @@ namespace WireMock.Net.OpenApiParser.Extensions
         {
             var mappings = new WireMockOpenApiParser().FromStream(stream, out diagnostic);
 
+            return WithMappingFromOpenApiStream(server, stream, null, out diagnostic);
+        }
+
+        /// <summary>
+        /// Register the mappings via an OpenAPI (swagger) V2 or V3 stream.
+        /// </summary>
+        /// <param name="server">The WireMockServer instance</param>
+        /// <param name="stream">Stream containing OpenAPI description to parse and use the mappings.</param>
+        /// <param name="settings">Additional settings</param>
+        /// <param name="diagnostic">Returns diagnostic object containing errors detected during parsing</param>
+        public static IWireMockServer WithMappingFromOpenApiStream(this IWireMockServer server, Stream stream, WireMockOpenApiParserSettings settings, out OpenApiDiagnostic diagnostic)
+        {
+            var mappings = new WireMockOpenApiParser().FromStream(stream, settings, out diagnostic);
+
             return server.WithMapping(mappings.ToArray());
         }
 
@@ -49,9 +73,10 @@ namespace WireMock.Net.OpenApiParser.Extensions
         /// </summary>
         /// <param name="server">The WireMockServer instance</param>
         /// <param name="document">The OpenAPI document to use as mappings.</param>
-        public static IWireMockServer WithMappingFromOpenApiDocument(this IWireMockServer server, OpenApiDocument document)
+        /// <param name="settings">Additional settings [optional]</param>
+        public static IWireMockServer WithMappingFromOpenApiDocument(this IWireMockServer server, OpenApiDocument document, WireMockOpenApiParserSettings settings = null)
         {
-            var mappings = new WireMockOpenApiParser().FromDocument(document);
+            var mappings = new WireMockOpenApiParser().FromDocument(document, settings);
 
             return server.WithMapping(mappings.ToArray());
         }
