+++ src/WireMock.Net.OpenApiParser/IWireMockOpenApiParser.cs
@@ -3,6 +3,7 @@ using System.IO;
 using Microsoft.OpenApi.Models;
 using Microsoft.OpenApi.Readers;
 using WireMock.Admin.Mappings;
+using WireMock.Net.OpenApiParser.Settings;
 
 namespace WireMock.Net.OpenApiParser
 {
@@ -12,11 +13,29 @@ namespace WireMock.Net.OpenApiParser
     public interface IWireMockOpenApiParser
     {
         /// <summary>
+        /// Generate <see cref="IEnumerable{MappingModel}"/> from a file-path.
+        /// </summary>
+        /// <param name="path">The path to read the OpenApi/Swagger/V2/V3 or Raml file.</param>
+        /// <param name="diagnostic">OpenApiDiagnostic output</param>
+        /// <returns>MappingModel</returns>
+        IEnumerable<MappingModel> FromFile(string path, out OpenApiDiagnostic diagnostic);
+
+        /// <summary>
+        /// Generate <see cref="IEnumerable{MappingModel}"/> from a file-path.
+        /// </summary>
+        /// <param name="path">The path to read the OpenApi/Swagger/V2/V3 or Raml file.</param>
+        /// <param name="settings">Additional settings</param>
+        /// <param name="diagnostic">OpenApiDiagnostic output</param>
+        /// <returns>MappingModel</returns>
+        IEnumerable<MappingModel> FromFile(string path, WireMockOpenApiParserSettings settings, out OpenApiDiagnostic diagnostic);
+
+        /// <summary>
         /// Generate <see cref="IEnumerable{MappingModel}"/> from an <seealso cref="OpenApiDocument"/>.
         /// </summary>
         /// <param name="document">The source OpenApiDocument</param>
+        /// <param name="settings">Additional settings [optional]</param>
         /// <returns>MappingModel</returns>
-        IEnumerable<MappingModel> FromDocument(OpenApiDocument document);
+        IEnumerable<MappingModel> FromDocument(OpenApiDocument document, WireMockOpenApiParserSettings settings = null);
 
         /// <summary>
         /// Generate <see cref="IEnumerable{MappingModel}"/> from a <seealso cref="Stream"/>.
@@ -27,11 +46,12 @@ namespace WireMock.Net.OpenApiParser
         IEnumerable<MappingModel> FromStream(Stream stream, out OpenApiDiagnostic diagnostic);
 
         /// <summary>
-        /// Generate <see cref="IEnumerable{MappingModel}"/> from a file-path.
+        /// Generate <see cref="IEnumerable{MappingModel}"/> from a <seealso cref="Stream"/>.
         /// </summary>
-        /// <param name="path">The path to read the OpenApi/Swagger/V2/V3 or Raml file.</param>
+        /// <param name="stream">The source stream</param>
+        /// <param name="settings">Additional settings</param>
         /// <param name="diagnostic">OpenApiDiagnostic output</param>
         /// <returns>MappingModel</returns>
-        IEnumerable<MappingModel> FromFile(string path, out OpenApiDiagnostic diagnostic);
+        IEnumerable<MappingModel> FromStream(Stream stream, WireMockOpenApiParserSettings settings, out OpenApiDiagnostic diagnostic);
     }
 }
